Fix additional stages, more cleanup, rename progress to retries

+++ libafl/src/corpus/testcase.rs
@@ -541,7 +541,7 @@ pub mod pybind {
         }
 
         #[getter]
-        fn executions(&self) -> usize {
+        fn executions(&self) -> u64 {
             *self.inner.as_ref().executions()
         }
